refactor(server): split shutdown.rs, which my own tests pushed over the limit
The warning-branch tests need a tracing capture layer and a paused clock, and
together they are longer than the shutdown sequence they cover — 589 lines,
which the Format job rejected, correctly.

Same treatment as `handler/messaging.rs` earlier in this cycle:
`shutdown/mod.rs` (166) holds the sequence, `shutdown/tests.rs` (435) holds the
suite. No behaviour change; all 12 shutdown tests still pass.

I pushed the previous commit without running `check_file_lengths.sh` and let CI
find this. Running the gate locally is the whole point of it existing.

Verified: 2,837 tests pass at all features, clippy clean with `-D warnings`,
fmt clean, all nine Format-job gates exit 0.
